👷 [+ci] CI Workaround (Change 1 of N)
This is a workaround for a bug discovered while fixing another bug.

1) SimpleSchema could not be built on the CI machine because the detached head state wasn't resolving to a mainline branch.

2) This fix had to happen in Common_Foundation, but there was a bug in its dependency scripts where it wasn't properly setting this repository's branch.

So, the fix is to temporarily disable the use of branches in SimpleSchema's setup, let that make it to SimpleSchema's main_stable branch, let Common_Foundation's main_stable cover the AutoSemVer issue (now that SimpleSchema won't be failing), then revert this change to allow SimpleSchema to continue.